+++ neuralprophet/components/seasonality/base.py
