+++ neuralprophet/components/trend/__init__.py
@@ -1,1 +1,1 @@
-from .trend import Trend  # noqa: F401
+from .base import Trend  # noqa: F401
