+++ neuralprophet/components/trend/base.py
